requirements.txt
Flask==3.0.0
gunicorn==21.2.0
numpy==1.26.4
pydub==0.25.1
scipy==1.13.0
pyloudnorm==0.1.1
google-cloud-storage==2.14.0
google-cloud-tasks==2.14.2
google-auth==2.22.0